address reviewer comments.
