+++ doc/content/media/validation/free_flow/heat_transfer/flat_plate/interface-temperature.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:3c45dd2ad0b3fc863ed75e6d7db0b58f74e411b3364c49cd1bf1e9e4783c5b85
+size 174135
