+++ doc/content/media/validation/free_flow/heat_transfer/flat_plate/vertical-temperature.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:497d2038325bff5ef446783b5e85ba0e039ed661a18a0d7f98c5a25dd803c036
+size 178761
