+++ doc/content/media/validation/free_flow/isothermal/2d_turbulent_channel/7_Ret395_dual_plots.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:bf52430cead360bd67829e7a86a821e6368a74373e8e73ab2ed9ea99fdd614d0
+size 37556
