+++ doc/content/media/validation/free_flow/isothermal/2d_turbulent_channel/8_Ret590_dual_plots.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:1c2e7958115a37115cb02e952ea4e7cd83aa8b2252400e97a025f1f14cf6874d
+size 32832
