+++ doc/content/media/validation/free_flow/isothermal/ercoftac_030_bfs/plots_cf_main.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:5f3284ccb90142b36cfb05f6609cfba31db51e2f0ac690790c4a5418e09b31c2
+size 224228
