+++ doc/content/media/validation/free_flow/isothermal/ercoftac_030_bfs/plots_cp_cf_error_main.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:874199e1d5bf3f8baceeb739daa9783564be35a1741ae0d75a5a7d1be5dc0daf
+size 355196
