+++ doc/content/media/validation/free_flow/isothermal/ercoftac_030_bfs/plots_cp_main.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:9ec199988828face442a481a1d2e89d86db7ec83934817daab5bcdfcae303fef
+size 230345
