+++ doc/content/media/validation/free_flow/isothermal/ercoftac_030_bfs/plots_u_profiles_error1_main.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:205145c8bc7fe0bafce607ada0ca93c0cd00aaa46b1f3410488b3a1d7e2b175e
+size 302857
