+++ doc/content/media/validation/free_flow/isothermal/ercoftac_030_bfs/plots_u_profiles_error2_main.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:3029f61e959aa07ab4095170ceb2f4122a570b09d3ec4fae0ed5fa398b73f6b9
+size 316562
