+++ doc/content/media/validation/free_flow/isothermal/ercoftac_030_bfs/plots_u_profiles_main.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:8fcdb8ef573f8f7cc68a2710b29c909dbac75bd1e1f5dfd2c80ecdf18b6ff850
+size 448846
