+++ doc/content/media/validation/free_flow/isothermal/vortex_shedding/cylinder/strouhal.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:1055e69d4482459cc4d73d66cbb4a97e26b243e4ca3e54bf229340d8ea7989f0
+size 22665
