+++ doc/content/media/validation/free_flow/jets/single-stream-jet/jet_CL_normal_velocity.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:3f659a34620478886f355287d6efaeba469ac527c2d0fc8677f0077934ee2b1f
+size 269259
