+++ doc/content/media/validation/free_flow/jets/single-stream-jet/jet_halfwidth.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:fa675057c3ac37e1fa3768b4a3ac6af2c55a395bd867a9578f1c99756b3e1cee
+size 198546
