+++ doc/content/media/validation/free_flow/jets/single-stream-jet/mesh.png
