+++ doc/content/media/validation/free_flow/jets/single-stream-jet/plot_results.py
@@ -1,11 +1,22 @@
 #!/usr/bin/env python3
 import os
-import numpy as np
 import pandas as pd
 import matplotlib.pyplot as plt
 
+SCRIPT_DIR = os.path.dirname(os.path.realpath(__file__))
+
+os.chdir(SCRIPT_DIR)
+
 D = 0.01
-REF = "reference"
+
+DATA = os.path.abspath(
+    os.path.join(
+        SCRIPT_DIR,
+        "../../../../../../../validation/free_flow/jets/single-stream-jet/RTL10000",
+    )
+)
+REF = os.path.join(DATA, "reference")
+GOLD = os.path.join(DATA, "gold")
 
 plt.rcParams["font.family"] = "serif"
 plt.rcParams["mathtext.fontset"] = "dejavuserif"
@@ -15,23 +26,27 @@ def load_ref(name):
     return pd.read_csv(os.path.join(REF, name), header=None, names=["x", "y"])
 
 
-def load_centerline_sim(fname):
-    df = pd.read_csv(fname).sort_values("z").reset_index(drop=True)
+def output_path(name):
+    current = os.path.join(DATA, name)
+    if os.path.exists(current):
+        return current
+    return os.path.join(GOLD, name)
+
+
+def load_centerline_sim(path):
+    df = pd.read_csv(path).sort_values("z").reset_index(drop=True)
     df["z_over_D"] = df["z"] / D
     df["Vin_over_Vc"] = df["vel_z"].iloc[0] / df["vel_z"]
     return df
 
 
-def load_halfwidth_sim(fname):
-    df = pd.read_csv(fname).sort_values("z").reset_index(drop=True)
+def load_halfwidth_sim(path):
+    df = pd.read_csv(path).sort_values("z").reset_index(drop=True)
     df["z_over_D"] = df["z"] / D
     df["rhalf_over_D"] = df["y_half"] / D
     return df
 
 
-# -------------------------
-# Load reference data
-# -------------------------
 exp_centerline = load_ref("exp_centerline.csv")
 default_centerline = load_ref("default_centerline.csv")
 calibrated_centerline = load_ref("calibrated_centerline.csv")
@@ -45,15 +60,11 @@ mi_halfwidth = load_ref("Mi_halfwidth.csv")
 calibrated_halfwidth = load_ref("calibrated_IP_RSM_halfwidth.csv")
 default_halfwidth = load_ref("default_IP_RSM_halfwidth.csv")
 
-# -------------------------
-# Load simulation data
-# -------------------------
-opgh_centerline = load_centerline_sim("jet_RTL10000_out_centerline_0002.csv")
-opgh_halfwidth = load_halfwidth_sim("combined_halfwidths.csv")
+opgh_centerline = load_centerline_sim(
+    output_path("jet_RTL10000_out_centerline_0002.csv")
+)
+opgh_halfwidth = load_halfwidth_sim(output_path("combined_halfwidths.csv"))
 
-# -------------------------
-# Centerline plot
-# -------------------------
 plt.figure()
 
 plt.plot(
@@ -73,7 +84,7 @@ plt.plot(
     opgh_centerline["Vin_over_Vc"],
     linestyle="-",
     color="red",
-    label="openPronghorn (RealizableTwoLayer, Wolfstein)",
+    label="OpenPronghorn (RealizableTwoLayer, Wolfstein)",
 )
 
 plt.plot(
@@ -82,7 +93,7 @@ plt.plot(
     linestyle="-",
     linewidth=2.0,
     color="blue",
-    label="default IP-RSM, Turutoglu et al. 2024",
+    label="Default IP-RSM, Turutoglu et al. 2024",
 )
 
 plt.plot(
@@ -91,7 +102,7 @@ plt.plot(
     linestyle="-",
     linewidth=2.0,
     color="orange",
-    label="calibrated IP-RSM, Turutoglu et al. 2024",
+    label="Calibrated IP-RSM, Turutoglu et al. 2024",
 )
 
 plt.plot(
@@ -100,7 +111,7 @@ plt.plot(
     linestyle="-",
     linewidth=2.0,
     color="green",
-    label="Panchapakesan & Lumley",
+    label="Panchapakesan & Lumley 1993",
 )
 
 plt.plot(
@@ -110,7 +121,7 @@ plt.plot(
     marker="x",
     markersize=7,
     color="purple",
-    label="Mi et al.",
+    label="Mi et al. 2013",
 )
 
 plt.plot(
@@ -134,11 +145,8 @@ plt.xticks(fontsize=14)
 plt.yticks(fontsize=14)
 plt.grid(True)
 plt.savefig("jet_CL_normal_velocity.png", dpi=300, bbox_inches="tight")
-plt.show()
+plt.close()
 
-# -------------------------
-# Half-width plot
-# -------------------------
 plt.figure()
 
 plt.plot(
@@ -160,7 +168,7 @@ plt.plot(
     marker="+",
     markersize=7,
     color="green",
-    label="Panchapakesan & Lumley correlation",
+    label="Panchapakesan & Lumley 1993",
 )
 
 plt.plot(
@@ -170,7 +178,7 @@ plt.plot(
     marker="x",
     markersize=7,
     color="purple",
-    label="Mi et al. correlation",
+    label="Mi et al. 2013",
 )
 
 plt.plot(
@@ -179,7 +187,7 @@ plt.plot(
     linestyle="-",
     linewidth=2.0,
     color="blue",
-    label="default IP-RSM, Turutoglu et al. 2024",
+    label="Default IP-RSM, Turutoglu et al. 2024",
 )
 
 plt.plot(
@@ -188,7 +196,7 @@ plt.plot(
     linestyle="-",
     linewidth=2.0,
     color="orange",
-    label="calibrated IP-RSM, Turutoglu et al. 2024",
+    label="Calibrated IP-RSM, Turutoglu et al. 2024",
 )
 
 plt.plot(
@@ -197,11 +205,12 @@ plt.plot(
     linestyle="-",
     linewidth=2.0,
     color="red",
-    label="openPronghorn (RealizableTwoLayer, Wolfstein)",
+    label="OpenPronghorn (RealizableTwoLayer, Wolfstein)",
 )
 
 plt.title(
-    "Jet half-width growth\nSingle-Stream Jet in Still Air, Re = 10000", fontsize=13
+    "Jet half-width growth\nSingle-Stream Jet in Still Air, Re = 10000",
+    fontsize=13,
 )
 plt.xlabel(r"$axial~location~[x/D]$", fontsize=14)
 plt.ylabel(r"$r_{1/2}/D~[-]$", fontsize=14)
@@ -212,4 +221,4 @@ plt.xticks(fontsize=14)
 plt.yticks(fontsize=14)
 plt.grid(True)
 plt.savefig("jet_halfwidth.png", dpi=300, bbox_inches="tight")
-plt.show()
+plt.close()
