+++ validation/free_flow/jets/single-stream-jet/RTL10000/gold/jet_RTL10000_out_centerline_0002.csv
@@ -1,0 +1,151 @@
+id,vel_z,x,y,z
+0,15.573947227008,0.0001,0.0001,0.001
+0.002,15.573688067565,0.0001,0.0001,0.003
+0.004,15.572995785706,0.0001,0.0001,0.005
+0.006,15.570203052126,0.0001,0.0001,0.007
+0.008,15.565520337858,0.0001,0.0001,0.009
+0.01,15.559625737435,0.0001,0.0001,0.011
+0.012,15.552626197452,0.0001,0.0001,0.013
+0.014,15.544435095404,0.0001,0.0001,0.015
+0.016,15.535028144038,0.0001,0.0001,0.017
+0.018,15.524421128285,0.0001,0.0001,0.019
+0.02,15.512596075928,0.0001,0.0001,0.021
+0.022,15.49949478776,0.0001,0.0001,0.023
+0.024,15.485052041866,0.0001,0.0001,0.025
+0.026,15.469204303152,0.0001,0.0001,0.027
+0.028,15.451867059354,0.0001,0.0001,0.029
+0.03,15.432922924673,0.0001,0.0001,0.031
+0.032,15.412233270776,0.0001,0.0001,0.033
+0.034,15.389645158357,0.0001,0.0001,0.035
+0.036,15.364985688199,0.0001,0.0001,0.037
+0.038,15.338051408693,0.0001,0.0001,0.039
+0.04,15.308603488346,0.0001,0.0001,0.041
+0.042,15.276368900557,0.0001,0.0001,0.043
+0.044,15.241039063525,0.0001,0.0001,0.045
+0.046,15.202263455462,0.0001,0.0001,0.047
+0.048,15.159641806787,0.0001,0.0001,0.049
+0.05,15.112719747118,0.0001,0.0001,0.051
+0.052,15.060987859678,0.0001,0.0001,0.053
+0.054,15.003880941355,0.0001,0.0001,0.055
+0.056,14.940777072559,0.0001,0.0001,0.057
+0.058,14.870999260633,0.0001,0.0001,0.059
+0.06,14.793823355046,0.0001,0.0001,0.061
+0.062,14.708493285716,0.0001,0.0001,0.063
+0.064,14.614241076963,0.0001,0.0001,0.065
+0.066,14.510313247798,0.0001,0.0001,0.067
+0.068,14.39600132935,0.0001,0.0001,0.069
+0.07,14.270680792949,0.0001,0.0001,0.071
+0.072,14.133853719977,0.0001,0.0001,0.073
+0.074,13.985190651455,0.0001,0.0001,0.075
+0.076,13.824568083938,0.0001,0.0001,0.077
+0.078,13.652097137282,0.0001,0.0001,0.079
+0.08,13.468139651701,0.0001,0.0001,0.081
+0.082,13.273310926634,0.0001,0.0001,0.083
+0.084,13.068467464409,0.0001,0.0001,0.085
+0.086,12.854689271473,0.0001,0.0001,0.087
+0.088,12.633234353503,0.0001,0.0001,0.089
+0.09,12.405495328409,0.0001,0.0001,0.091
+0.092,12.17294858189,0.0001,0.0001,0.093
+0.094,11.93709359851,0.0001,0.0001,0.095
+0.096,11.699403195092,0.0001,0.0001,0.097
+0.098,11.461272506213,0.0001,0.0001,0.099
+0.1,11.223976011585,0.0001,0.0001,0.101
+0.102,10.988635085021,0.0001,0.0001,0.103
+0.104,10.756204092407,0.0001,0.0001,0.105
+0.106,10.527464347557,0.0001,0.0001,0.107
+0.108,10.303023508913,0.0001,0.0001,0.109
+0.11,10.083333642425,0.0001,0.0001,0.111
+0.112,9.8687183902589,0.0001,0.0001,0.113
+0.114,9.6593930990376,0.0001,0.0001,0.115
+0.116,9.4554881923592,0.0001,0.0001,0.117
+0.118,9.2570694246757,0.0001,0.0001,0.119
+0.12,9.064155349437,0.0001,0.0001,0.121
+0.122,8.8767323833396,0.0001,0.0001,0.123
+0.124,8.6947667191238,0.0001,0.0001,0.125
+0.126,8.5182133407783,0.0001,0.0001,0.127
+0.128,8.3470198906856,0.0001,0.0001,0.129
+0.13,8.1811295757834,0.0001,0.0001,0.131
+0.132,8.0204838813187,0.0001,0.0001,0.133
+0.134,7.8650284193145,0.0001,0.0001,0.135
+0.136,7.7146810506294,0.0001,0.0001,0.137
+0.138,7.5693582415775,0.0001,0.0001,0.139
+0.14,7.4289653003754,0.0001,0.0001,0.141
+0.142,7.2933966748483,0.0001,0.0001,0.143
+0.144,7.1625361582939,0.0001,0.0001,0.145
+0.146,7.0362583561336,0.0001,0.0001,0.147
+0.148,6.9144305197631,0.0001,0.0001,0.149
+0.15,6.7969142158803,0.0001,0.0001,0.151
+0.152,6.6835682644674,0.0001,0.0001,0.153
+0.154,6.5742513567461,0.0001,0.0001,0.155
+0.156,6.4688208727323,0.0001,0.0001,0.157
+0.158,6.3671760188294,0.0001,0.0001,0.159
+0.16,6.2690905768594,0.0001,0.0001,0.161
+0.162,6.1744229950385,0.0001,0.0001,0.163
+0.164,6.0830602767354,0.0001,0.0001,0.165
+0.166,5.9948639099772,0.0001,0.0001,0.167
+0.168,5.9097013326581,0.0001,0.0001,0.169
+0.17,5.8274528232942,0.0001,0.0001,0.171
+0.172,5.7479935605628,0.0001,0.0001,0.173
+0.174,5.6711997889242,0.0001,0.0001,0.175
+0.176,5.5969505875085,0.0001,0.0001,0.177
+0.178,5.5251208596497,0.0001,0.0001,0.179
+0.18,5.4556467765188,0.0001,0.0001,0.181
+0.182,5.3884003748644,0.0001,0.0001,0.183
+0.184,5.3232985315522,0.0001,0.0001,0.185
+0.186,5.2602459170363,0.0001,0.0001,0.187
+0.188,5.1991887364702,0.0001,0.0001,0.189
+0.19,5.1400253008202,0.0001,0.0001,0.191
+0.192,5.0828619827624,0.0001,0.0001,0.193
+0.194,5.026845652919,0.0001,0.0001,0.195
+0.196,4.9723469815651,0.0001,0.0001,0.197
+0.198,4.9193546548479,0.0001,0.0001,0.199
+0.2,4.8678148899647,0.0001,0.0001,0.201
+0.202,4.8176492528924,0.0001,0.0001,0.203
+0.204,4.7688034242022,0.0001,0.0001,0.205
+0.206,4.7211806973574,0.0001,0.0001,0.207
+0.208,4.6747297572039,0.0001,0.0001,0.209
+0.21,4.6293575953003,0.0001,0.0001,0.211
+0.212,4.584984422937,0.0001,0.0001,0.213
+0.214,4.5415399246007,0.0001,0.0001,0.215
+0.216,4.4989599536341,0.0001,0.0001,0.217
+0.218,4.4571874781572,0.0001,0.0001,0.219
+0.22,4.4161573949609,0.0001,0.0001,0.221
+0.222,4.3758128406576,0.0001,0.0001,0.223
+0.224,4.3361103746785,0.0001,0.0001,0.225
+0.226,4.2970058020385,0.0001,0.0001,0.227
+0.228,4.2585554348308,0.0001,0.0001,0.229
+0.23,4.2202812775263,0.0001,0.0001,0.231
+0.232,4.1825193186485,0.0001,0.0001,0.233
+0.234,4.1451325228809,0.0001,0.0001,0.235
+0.236,4.1080710679674,0.0001,0.0001,0.237
+0.238,4.071329447726,0.0001,0.0001,0.239
+0.24,4.0348720188577,0.0001,0.0001,0.241
+0.242,3.9986843370383,0.0001,0.0001,0.243
+0.244,3.9627131841699,0.0001,0.0001,0.245
+0.246,3.9269606999335,0.0001,0.0001,0.247
+0.248,3.8913996334743,0.0001,0.0001,0.249
+0.25,3.856010577559,0.0001,0.0001,0.251
+0.252,3.820784047616,0.0001,0.0001,0.253
+0.254,3.7857156686941,0.0001,0.0001,0.255
+0.256,3.7508057667219,0.0001,0.0001,0.257
+0.258,3.7160557296896,0.0001,0.0001,0.259
+0.26,3.6814698451988,0.0001,0.0001,0.261
+0.262,3.6470541858403,0.0001,0.0001,0.263
+0.264,3.6128299969647,0.0001,0.0001,0.265
+0.266,3.5788107968636,0.0001,0.0001,0.267
+0.268,3.5450899670894,0.0001,0.0001,0.269
+0.27,3.5116843418492,0.0001,0.0001,0.271
+0.272,3.478586545818,0.0001,0.0001,0.273
+0.274,3.4458426554784,0.0001,0.0001,0.275
+0.276,3.4134963483418,0.0001,0.0001,0.277
+0.278,3.3815688552559,0.0001,0.0001,0.279
+0.28,3.3501100289078,0.0001,0.0001,0.281
+0.282,3.3191654294174,0.0001,0.0001,0.283
+0.284,3.2888360884882,0.0001,0.0001,0.285
+0.286,3.2591341273739,0.0001,0.0001,0.287
+0.288,3.2303869791841,0.0001,0.0001,0.289
+0.29,3.2022038641269,0.0001,0.0001,0.291
+0.292,3.1762072540076,0.0001,0.0001,0.293
+0.294,3.1491813131898,0.0001,0.0001,0.295
+0.296,3.1300764450208,0.0001,0.0001,0.297
+0.298,3.100147215917,0.0001,0.0001,0.299
